refactor(LottosController) : 로또 구매를 담당하는 컨트롤러 생성
