+++ src/main/java/lotto/Application.java
@@ -1,11 +1,11 @@
 package lotto;
 
-import lotto.controller.LottoController;
+import lotto.controller.Controller;
 
 public class Application {
 
     public static void main(String[] args) {
-        LottoController lottoController = new LottoController();
-        lottoController.play();
+        Controller controller = new Controller();
+        controller.play();
     }
 }
