+++ src/main/java/lotto/controller/LottosController.java
@@ -1,27 +1,16 @@
 package lotto.controller;
 
+import lotto.domain.Money;
+import lotto.domain.factory.LottoFactory;
 import lotto.domain.factory.MoneyFactory;
 import lotto.domain.lotto.Lotto;
-import lotto.domain.factory.LottoFactory;
-import lotto.domain.factory.NumberFactory;
-import lotto.domain.lotto.WinningLotto;
 import lotto.domain.lotto.Lottos;
-import lotto.domain.Money;
-import lotto.domain.lotto.Number;
-import lotto.domain.Result;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
-public class LottoController {
+public class LottosController {
 
-    public void play() {
-        // 로또 구매
-        Lottos lottos = buyLottos();
-        // 결과 출력
-        printResult(lottos);
-    }
-
-    private Lottos buyLottos() {
+    public Lottos buyLottos() {
         Money money = requestMoney();
         int manualLottoCount = requestManualLottoCount(money);
         int autoLottoCount = money.countToBuyLotto();
@@ -81,39 +70,6 @@ public class LottoController {
     private void buyAutoLottos(int autoLottoCount, Lottos lottos) {
         while (autoLottoCount-- > 0) {
             lottos.add(LottoFactory.auto());
-        }
-    }
-
-    private Lotto requestLotto() {
-        try {
-            String input = InputView.inputLastWeekWinningNumbers();
-            return LottoFactory.valueOf(input);
-        } catch (IllegalArgumentException exception) {
-            OutputView.printException(exception);
-            return requestLotto();
-        }
-    }
-
-    private void printResult(Lottos lottos) {
-        Money totalMoney = lottos.totalPrice();
-        Result result = lottos.getResult(requestWinningLotto());
-        OutputView.printResult(result);
-        OutputView.printRateOfProfit(totalMoney.getRateOfProfit(result.getTotalProfit()));
-    }
-
-    private WinningLotto requestWinningLotto() {
-        Lotto lotto = requestLotto();
-        return requestWinningLottoContainingBonusNumber(lotto);
-    }
-
-    private WinningLotto requestWinningLottoContainingBonusNumber(Lotto lotto) {
-        try {
-            String input = InputView.inputBonusNumber();
-            Number bonusNumber = NumberFactory.valueOf(input);
-            return new WinningLotto(lotto, bonusNumber);
-        } catch (IllegalArgumentException exception) {
-            OutputView.printException(exception);
-            return requestWinningLottoContainingBonusNumber(lotto);
         }
     }
 }
